requirements.txt
elasticsearch>=8.0.0
sentence-transformers>=2.2.0
pandas>=1.5.0
numpy>=1.21.0
tqdm>=4.64.0
python-dotenv>=0.21.0
requests>=2.28.0
